+++ src/components/Activity/ActivityMintSuccessDialog.tsx
@@ -6,7 +6,7 @@ import { ChainId } from '@masknet/web3-shared-evm';
 import SuccessShieldIcon from '@/assets/success-shield.svg';
 import { useActivityCompose } from '@/components/Activity/hooks/useActivityCompose.js';
 import { Link } from '@/components/Activity/Link.js';
-import { CloseButton } from '@/components/CloseButton.js';
+import { CloseButton } from '@/components/IconButton.js';
 import { Modal } from '@/components/Modal.js';
 import { Popover } from '@/components/Popover.js';
 import type { Chars } from '@/helpers/chars.js';
