+++ src/components/SolanaAccountModal.tsx
@@ -6,7 +6,7 @@ import { useQuery } from '@tanstack/react-query';
 import { forwardRef } from 'react';
 
 import { ClickableButton } from '@/components/ClickableButton.js';
-import { CloseButton } from '@/components/CloseButton.js';
+import { CloseButton } from '@/components/IconButton.js';
 import { CopyTextButton } from '@/components/CopyTextButton.js';
 import { Image } from '@/components/Image.js';
 import { Modal } from '@/components/Modal.js';
