+++ src/components/Tips/TipsModalHeader.tsx
@@ -4,7 +4,7 @@ import { rootRouteId, useMatch } from '@tanstack/react-router';
 
 import LeftArrowIcon from '@/assets/left-arrow.svg';
 import { ClickableButton } from '@/components/ClickableButton.js';
-import { CloseButton } from '@/components/CloseButton.js';
+import { CloseButton } from '@/components/IconButton.js';
 import { router, TipsRoutePath } from '@/components/Tips/TipsModalRouter.js';
 import { Tooltip } from '@/components/Tooltip.js';
 import { useIsSmall } from '@/hooks/useMediaQuery.js';
