+++ src/modals/ConnectWalletModal.tsx
@@ -6,7 +6,7 @@ import { forwardRef, memo, type ReactNode } from 'react';
 
 import LoadingIcon from '@/assets/loading.svg';
 import { ClickableButton } from '@/components/ClickableButton.js';
-import { CloseButton } from '@/components/CloseButton.js';
+import { CloseButton } from '@/components/IconButton.js';
 import { Image } from '@/components/Image.js';
 import { Modal } from '@/components/Modal.js';
 import { Popover } from '@/components/Popover.js';
